pyproject.toml
[tool.poetry]
name = "lab_1"
description = "Training pipeline for CIFAR-100 dataset"
authors = ["Oleksandr Klymenchuk <klimencuks@gmail.com>"]
version = "0.01"

[tool.poetry.dependencies]
python = "~3.10"
torch = { version = "2.1.0+cu118", source = "pytorch" }
torchvision = { version = "0.16.0+cu118", source = "pytorch" }
tqdm = "4.64.1"
matplotlib = "3.6.3"
numpy = "1.22.4"
pyyaml = "6.0"
scipy = "1.13.0rc1"

[tool.poetry.dev-dependencies]
mypy = "0.991"
ruff = "0.0.254"
black = "23.1.0"
isort = "5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
priority = "supplemental"
url = "https://download.pytorch.org/whl/cu118"
